requirements.txt
flask
flask-cors
psycopg2-binary
scikit-learn
pandas
numpy
python-dotenv
requests